// src/core/integrations/oauth/type.ts
// ---------- Google Oauth Types ----------

export type GoogleOauthApiResponse = {
    access_token: string;
    expires_in: number;
    refresh_token: string;
    scope: string;
    token_type: string;
    id_token: string;
}

export type GoogleUserInfoResponse = {
    sub: string;
    name: string;
    given_name: string;
    picture: string;
    email: string;
    email_verified: boolean;
}


// ---------- Github Oauth Types ----------

export type GithubOauthApiResponse = {
    access_token: string;
    token_type: string;
    scope: string;
}

export type GitHubUserResponse = {
    login: string;
    id: number;
    node_id: string;

    avatar_url: string;
    gravatar_id: string;
    url: string;
    html_url: string;

    followers_url: string;
    following_url: string;
    gists_url: string;
    starred_url: string;
    subscriptions_url: string;
    organizations_url: string;
    repos_url: string;
    events_url: string;
    received_events_url: string;

    type: "User" | "Organization";
    user_view_type?: "public" | "private";
    site_admin: boolean;

    name: string | null;
    company: string | null;
    blog: string;
    location: string | null;
    email: string | null;
    hireable: boolean | null;
    bio: string | null;
    twitter_username: string | null;
    notification_email: string | null;

    public_repos: number;
    public_gists: number;
    followers: number;
    following: number;

    created_at: string;
    updated_at: string;

    private_gists?: number;
    total_private_repos?: number;
    owned_private_repos?: number;
    disk_usage?: number;
    collaborators?: number;
    two_factor_authentication?: boolean;

    plan?: {
        name: string;
        space: number;
        collaborators: number;
        private_repos: number;
    };
};

export type GitHubEmailResponse = {
    email: string;
    primary: boolean;
    verified: boolean;
    visibility: "public" | "private" | null;
}